requirements.txt
absl-py==1.0.0
aiohttp @ file:///tmp/build/80754af9/aiohttp_1637857092750/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
asynctest==0.13.0
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
black==21.4b2
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi @ file:///tmp/build/80754af9/cffi_1636541929935/work
charset-normalizer==2.0.12
click @ file:///tmp/build/80754af9/click_1646056591685/work
cloudpickle==2.0.0
cryptography @ file:///tmp/build/80754af9/cryptography_1639414570729/work
cycler==0.11.0
Cython==0.29.28
detectron2==0.6+cpu
elastic-transport==8.0.1
elasticsearch==8.0.1
Flask @ file:///tmp/build/80754af9/flask_1634118196080/work
fonttools==4.29.1
frozenlist @ file:///tmp/build/80754af9/frozenlist_1637767116045/work
future==0.18.2
fvcore==0.1.5.post20220212
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h5py==3.6.0
hydra-core==1.1.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.16.0
importlib-metadata==4.11.1
importlib-resources==5.4.0
iopath==0.1.9
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
multidict @ file:///opt/conda/conda-bld/multidict_1640703752579/work
mypy-extensions==0.4.3
networkx==2.6.3
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634106693478/work
oauthlib==3.2.0
olefile==0.46
omegaconf==2.1.1
opencv-python==4.5.5.62
packaging==21.3
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
pathspec==0.9.0
Pillow==8.4.0
portalocker==2.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydot==1.4.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.7
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
python-dotenv==0.19.2
PyWavelets==1.2.0
PyYAML==6.0
regex==2022.1.18
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scipy==1.7.3
Shapely==1.8.1.post1
six @ file:///tmp/build/80754af9/six_1644875935023/work
submitit==1.4.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
tifffile==2021.11.2
timm==0.5.4
toml==0.10.2
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tqdm==4.62.3
typed-ast==1.5.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
yacs==0.1.8
yarl @ file:///tmp/build/80754af9/yarl_1606939952652/work
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
